requirements.txt

langchain==0.3.21
langchain-core==0.3.45
langchain-openai==0.3.7
langchain-text-splitters==0.3.7
langchain-community==0.3.20
langgraph==0.3.15

llama-index==0.12.21
llama-index-vector-stores-milvus==0.3.1
llama-index-retrievers-bm25==0.5.2
llama-index-embeddings-huggingface==0.5.5

sentence-transformers==3.4.1
transformers==4.50.3
accelerate==1.13.0
torch==2.6.0
torchvision==0.21.0
open_clip_torch==2.30.0

pymupdf==1.25.3
unstructured[all-docs]==0.17.2
paddlepaddle==3.1.0
paddleocr==2.10.0
python-multipart==0.0.16
pillow==12.0.0
opencv-python==4.11.0.86

python-dotenv==1.1.0
tiktoken==0.9.0
numpy==2.2.3
pandas==2.3.3
typing_extensions==4.12.2
chardet==5.2.0
